Cleaned up build process, got rid of random SPDLog warnings.
Basically, the CMake based build process is a lot cleaner. And this should allow creating games/other test projects with the engine much easier. However, I am really not sure how exactly this will work in terms of dependencies and such. Seems to be able to build purely based on the GLEngine and CPPModelLibrary static libs, but this may just be within the given CMake context.